# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Grow.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/plants": {
      "post": {
        "tags": [
          "Plants"
        ],
        "operationId": "CreatePlant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePlantResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Plants"
        ],
        "operationId": "SearchPlants",
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlantDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/{plantId}/events": {
      "post": {
        "tags": [
          "Plants"
        ],
        "operationId": "AddPlantEvent",
        "parameters": [
          {
            "name": "plantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddEventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/{plantId}/groups/{plantGroupId}": {
      "post": {
        "tags": [
          "Plants"
        ],
        "operationId": "AddPlantToGroup",
        "parameters": [
          {
            "name": "plantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "plantGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Plants"
        ],
        "operationId": "RemovePlantFromGroup",
        "parameters": [
          {
            "name": "plantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "plantGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/plant-groups": {
      "post": {
        "tags": [
          "Plant Groups"
        ],
        "operationId": "CreatePlantGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlantGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePlantGroupResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Plant Groups"
        ],
        "operationId": "GetPlantGroups",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "searchName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlantGroupDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/species": {
      "get": {
        "tags": [
          "Species"
        ],
        "operationId": "GetSpecies",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "searchName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecieDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Species"
        ],
        "operationId": "CreateSpecie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpecieRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSpecieResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/species/{specieId}/interval/{actionType}": {
      "post": {
        "tags": [
          "Species"
        ],
        "operationId": "UpdateSpecieInterval",
        "parameters": [
          {
            "name": "specieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "actionType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntervalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/me": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "GetMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/csrf": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "GetCsrfToken",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfTokenResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddEventRequest": {
        "required": [
          "type",
          "executedAt"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/PlantActionType"
          },
          "executedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AddEventResponse": {
        "required": [
          "plantEventId"
        ],
        "type": "object",
        "properties": {
          "plantEventId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreatePlantGroupRequest": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/GroupType"
          }
        }
      },
      "CreatePlantGroupResponse": {
        "required": [
          "createdPlantGroupId"
        ],
        "type": "object",
        "properties": {
          "createdPlantGroupId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreatePlantRequest": {
        "required": [
          "customId",
          "specieId"
        ],
        "type": "object",
        "properties": {
          "customId": {
            "type": "string"
          },
          "specieId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreatePlantResponse": {
        "required": [
          "createdPlantId"
        ],
        "type": "object",
        "properties": {
          "createdPlantId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateSpecieRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateSpecieResponse": {
        "required": [
          "specieId"
        ],
        "type": "object",
        "properties": {
          "specieId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateUserRequest": {
        "required": [
          "email",
          "username"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "CsrfTokenResponse": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "GroupType": {
        "type": "integer"
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "twoFactorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "twoFactorRecoveryCode": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "MeResponse": {
        "required": [
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
      "PlantActionType": {
        "type": "integer"
      },
      "PlantDto": {
        "required": [
          "id",
          "customId",
          "specieId",
          "plantGroupIds",
          "lastExecutions",
          "nextDates"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customId": {
            "type": "string"
          },
          "specieId": {
            "type": "string",
            "format": "uuid"
          },
          "plantGroupIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "lastExecutions": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "format": "date-time"
            }
          },
          "nextDates": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "format": "date"
            }
          }
        }
      },
      "PlantGroupDto": {
        "required": [
          "id",
          "name",
          "type",
          "plantIds"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/GroupType"
          },
          "plantIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "SpecieDto": {
        "required": [
          "id",
          "name",
          "intervals"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "intervals": {
            "type": "object",
            "additionalProperties": {
              "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
              "type": "string"
            }
          }
        }
      },
      "UpdateIntervalRequest": {
        "required": [
          "interval"
        ],
        "type": "object",
        "properties": {
          "interval": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Plants"
    },
    {
      "name": "Plant Groups"
    },
    {
      "name": "Species"
    },
    {
      "name": "Users"
    }
  ]
}